agregado
cambios agregado

+++ src/main/java/py/com/sgipy/miesys/controllers/GeneroJpaController.java
@@ -6,13 +6,16 @@
 package py.com.sgipy.miesys.controllers;
 
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import py.com.sgipy.miesys.entities.Persona;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import py.com.sgipy.miesys.controllers.exceptions.IllegalOrphanException;
 import py.com.sgipy.miesys.controllers.exceptions.NonexistentEntityException;
 import py.com.sgipy.miesys.entities.Genero;
 
@@ -32,11 +35,29 @@ public class GeneroJpaController implements Serializable {
     }
 
     public void create(Genero genero) {
+        if (genero.getPersonaList() == null) {
+            genero.setPersonaList(new ArrayList<Persona>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Persona> attachedPersonaList = new ArrayList<Persona>();
+            for (Persona personaListPersonaToAttach : genero.getPersonaList()) {
+                personaListPersonaToAttach = em.getReference(personaListPersonaToAttach.getClass(), personaListPersonaToAttach.getPersona());
+                attachedPersonaList.add(personaListPersonaToAttach);
+            }
+            genero.setPersonaList(attachedPersonaList);
             em.persist(genero);
+            for (Persona personaListPersona : genero.getPersonaList()) {
+                Genero oldGeneroOfPersonaListPersona = personaListPersona.getGenero();
+                personaListPersona.setGenero(genero);
+                personaListPersona = em.merge(personaListPersona);
+                if (oldGeneroOfPersonaListPersona != null) {
+                    oldGeneroOfPersonaListPersona.getPersonaList().remove(personaListPersona);
+                    oldGeneroOfPersonaListPersona = em.merge(oldGeneroOfPersonaListPersona);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -45,12 +66,45 @@ public class GeneroJpaController implements Serializable {
         }
     }
 
-    public void edit(Genero genero) throws NonexistentEntityException, Exception {
+    public void edit(Genero genero) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Genero persistentGenero = em.find(Genero.class, genero.getGenero());
+            List<Persona> personaListOld = persistentGenero.getPersonaList();
+            List<Persona> personaListNew = genero.getPersonaList();
+            List<String> illegalOrphanMessages = null;
+            for (Persona personaListOldPersona : personaListOld) {
+                if (!personaListNew.contains(personaListOldPersona)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Persona " + personaListOldPersona + " since its genero field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Persona> attachedPersonaListNew = new ArrayList<Persona>();
+            for (Persona personaListNewPersonaToAttach : personaListNew) {
+                personaListNewPersonaToAttach = em.getReference(personaListNewPersonaToAttach.getClass(), personaListNewPersonaToAttach.getPersona());
+                attachedPersonaListNew.add(personaListNewPersonaToAttach);
+            }
+            personaListNew = attachedPersonaListNew;
+            genero.setPersonaList(personaListNew);
             genero = em.merge(genero);
+            for (Persona personaListNewPersona : personaListNew) {
+                if (!personaListOld.contains(personaListNewPersona)) {
+                    Genero oldGeneroOfPersonaListNewPersona = personaListNewPersona.getGenero();
+                    personaListNewPersona.setGenero(genero);
+                    personaListNewPersona = em.merge(personaListNewPersona);
+                    if (oldGeneroOfPersonaListNewPersona != null && !oldGeneroOfPersonaListNewPersona.equals(genero)) {
+                        oldGeneroOfPersonaListNewPersona.getPersonaList().remove(personaListNewPersona);
+                        oldGeneroOfPersonaListNewPersona = em.merge(oldGeneroOfPersonaListNewPersona);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -68,7 +122,7 @@ public class GeneroJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -79,6 +133,17 @@ public class GeneroJpaController implements Serializable {
                 genero.getGenero();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The genero with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Persona> personaListOrphanCheck = genero.getPersonaList();
+            for (Persona personaListOrphanCheckPersona : personaListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Genero (" + genero + ") cannot be destroyed since the Persona " + personaListOrphanCheckPersona + " in its personaList field has a non-nullable genero field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(genero);
             em.getTransaction().commit();
